全てmain.goに書いてみた
